Add files via upload
This project is about analyzing the price data of Bitcoin and Ethereum using Python and some popular libraries like Pandas, Numpy, and Plotly. The aim of the project is to explore the data and identify any trends or patterns in the price of these cryptocurrencies over a certain period of time.